refactor(ui): InputValidator 클래스를 ui 패키지로 이동
입력 검증 기능이 사용자 입력과 직관적으로 연관되어 있어 ui 패키지로 이동
